requirements.txt
python-telegram-bot>=21,<23
httpx>=0.27,<0.29
python-dotenv>=1.0,<2.0
